Extract helpers to utils
