+++ extensions/sql-migration/src/wizard/integrationRuntimePage.ts
@@ -10,12 +10,11 @@ import { MigrationMode, MigrationStateModel, NetworkContainerType, StateChangeEv
 import { CreateSqlMigrationServiceDialog } from '../dialog/createSqlMigrationService/createSqlMigrationServiceDialog';
 import * as constants from '../constants/strings';
 import { WIZARD_INPUT_COMPONENT_WIDTH } from './wizardController';
-import { getFullResourceGroupFromId, getSqlMigrationService, getSqlMigrationServiceAuthKeys, getSqlMigrationServiceMonitoringData, regenerateSqlMigrationServiceAuthKey, SqlMigrationService, SqlVMServer } from '../api/azure';
+import { getFullResourceGroupFromId, getSqlMigrationService, getSqlMigrationServiceMonitoringData, SqlVMServer } from '../api/azure';
 import { IconPathHelper } from '../constants/iconPathHelper';
 import { logError, TelemetryViews } from '../telemetry';
 import * as utils from '../api/utils';
 import * as styles from '../constants/styles';
-import { azureResource } from 'azurecore';
 
 export class IntergrationRuntimePage extends MigrationWizardPage {
 	private _view!: azdata.ModelView;
@@ -504,9 +503,9 @@ export class IntergrationRuntimePage extends MigrationWizardPage {
 				CSSStyles: { ...styles.BODY_CSS }
 			}).component();
 
-		const instructions = createRegistrationInstructions(this._view, false);
+		const instructions = utils.createRegistrationInstructions(this._view, false);
 
-		this._authKeyTable = createAuthenticationKeyTable(this._view, '50px', '500px');
+		this._authKeyTable = utils.createAuthenticationKeyTable(this._view, '50px', '500px');
 
 		statusContainer.addItems([
 			this._dmsStatusInfoBox,
@@ -654,8 +653,7 @@ export class IntergrationRuntimePage extends MigrationWizardPage {
 				// exit if new call has started
 				if (callSequence !== this._lastIn) { return; }
 
-				// await this._authKeyTable.setDataValues(data);
-				await refreshAuthenticationKeyTable(this._view, this._authKeyTable, account, subscription, resourceGroup, location, migrationService);
+				await utils.refreshAuthenticationKeyTable(this._view, this._authKeyTable, account, subscription, resourceGroup, location, migrationService);
 
 				this.migrationStateModel._sqlMigrationService = migrationService;
 				this.migrationStateModel._sqlMigrationServiceSubscription = subscription;
@@ -674,209 +672,4 @@ export class IntergrationRuntimePage extends MigrationWizardPage {
 			}
 		}
 	}
-}
-
-export function createAuthenticationKeyTable(view: azdata.ModelView, columnWidth: string, stretchWidth: string): azdata.DeclarativeTableComponent {
-	const authKeyTable = view.modelBuilder.declarativeTable()
-		.withProps({
-			ariaLabel: constants.DATABASE_MIGRATION_SERVICE_AUTHENTICATION_KEYS,
-			columns: [
-				{
-					displayName: constants.NAME,
-					valueType: azdata.DeclarativeDataType.string,
-					width: columnWidth,
-					isReadOnly: true,
-					rowCssStyles: { ...styles.BODY_CSS },
-					headerCssStyles: { ...styles.BODY_CSS, 'font-weight': '600' }
-				},
-				{
-					displayName: constants.AUTH_KEY_COLUMN_HEADER,
-					valueType: azdata.DeclarativeDataType.string,
-					width: stretchWidth,
-					isReadOnly: true,
-					rowCssStyles: { ...styles.BODY_CSS },
-					headerCssStyles: { ...styles.BODY_CSS, 'font-weight': '600' }
-				},
-				{
-					displayName: '',
-					valueType: azdata.DeclarativeDataType.component,
-					width: columnWidth,
-					isReadOnly: true,
-					rowCssStyles: { ...styles.BODY_CSS },
-					headerCssStyles: { ...styles.BODY_CSS }
-				}
-			],
-			CSSStyles: { 'margin-top': '5px', 'width': WIZARD_INPUT_COMPONENT_WIDTH }
-		}).component();
-	return authKeyTable;
-}
-
-export async function refreshAuthenticationKeyTable(view: azdata.ModelView, table: azdata.DeclarativeTableComponent, account: azdata.Account, subscription: azureResource.AzureResourceSubscription, resourceGroup: string, location: string, service: SqlMigrationService): Promise<void> {
-	var _disposables: vscode.Disposable[] = [];
-
-	const copyKey1Button = view.modelBuilder.button().withProps({
-		title: constants.COPY_KEY1,
-		iconPath: IconPathHelper.copy,
-		ariaLabel: constants.COPY_KEY1,
-	}).component();
-
-	_disposables.push(copyKey1Button.onDidClick(async (e) => {
-		await vscode.env.clipboard.writeText(<string>table.dataValues![0][1].value);
-		void vscode.window.showInformationMessage(constants.SERVICE_KEY1_COPIED_HELP);
-	}));
-
-	const copyKey2Button = view.modelBuilder.button().withProps({
-		title: constants.COPY_KEY2,
-		iconPath: IconPathHelper.copy,
-		ariaLabel: constants.COPY_KEY2,
-	}).component();
-
-	_disposables.push(copyKey2Button.onDidClick(async (e) => {
-		await vscode.env.clipboard.writeText(<string>table.dataValues![1][1].value);
-		void vscode.window.showInformationMessage(constants.SERVICE_KEY2_COPIED_HELP);
-	}));
-
-	const refreshKey1Button = view.modelBuilder.button().withProps({
-		title: constants.REFRESH_KEY1,
-		iconPath: IconPathHelper.refresh,
-		ariaLabel: constants.REFRESH_KEY1,
-	}).component();
-
-	_disposables.push(refreshKey1Button.onDidClick(async (e) => {
-		const keys = await regenerateSqlMigrationServiceAuthKey(
-			account,
-			subscription,
-			resourceGroup,
-			location,
-			service.name,
-			'authKey1');
-
-		const dataValues = table.dataValues!;
-		dataValues![0][1].value = keys.authKey1;
-		await table.setDataValues([]);
-		await table.setDataValues(dataValues);
-		await vscode.window.showInformationMessage(constants.AUTH_KEY_REFRESHED(constants.SERVICE_KEY1_LABEL));
-	}));
-
-	const refreshKey2Button = view.modelBuilder.button().withProps({
-		title: constants.REFRESH_KEY2,
-		iconPath: IconPathHelper.refresh,
-		ariaLabel: constants.REFRESH_KEY2,
-	}).component();
-
-	_disposables.push(refreshKey2Button.onDidClick(async (e) => {
-		const keys = await regenerateSqlMigrationServiceAuthKey(
-			account,
-			subscription,
-			resourceGroup,
-			location,
-			service.name,
-			'authKey2');
-
-		const dataValues = table.dataValues!;
-		dataValues![1][1].value = keys.authKey2;
-		await table.setDataValues([]);
-		await table.setDataValues(dataValues);
-		await vscode.window.showInformationMessage(constants.AUTH_KEY_REFRESHED(constants.SERVICE_KEY2_LABEL));
-	}));
-
-	const keys = await getSqlMigrationServiceAuthKeys(
-		account,
-		subscription,
-		resourceGroup,
-		location,
-		service.name);
-
-
-	await table.updateProperties({
-		dataValues: [
-			[
-				{
-					value: constants.SERVICE_KEY1_LABEL
-				},
-				{
-					value: keys.authKey1
-				},
-				{
-					value: view.modelBuilder.flexContainer().withItems([copyKey1Button, refreshKey1Button]).component()
-				}
-			],
-			[
-				{
-					value: constants.SERVICE_KEY2_LABEL
-				},
-				{
-					value: keys.authKey2
-				},
-				{
-					value: view.modelBuilder.flexContainer().withItems([copyKey2Button, refreshKey2Button]).component()
-				}
-			]
-		]
-	});
-}
-
-export function createRegistrationInstructions(view: azdata.ModelView, testConnectionButton: boolean): azdata.FlexContainer {
-	const setupIRHeadingText = view.modelBuilder.text().withProps({
-		value: constants.SERVICE_CONTAINER_HEADING,
-		CSSStyles: {
-			...styles.LABEL_CSS
-		}
-	}).component();
-
-	const setupIRdescription1 = view.modelBuilder.text().withProps({
-		value: constants.SERVICE_CONTAINER_DESCRIPTION1,
-		CSSStyles: {
-			...styles.BODY_CSS
-		}
-	}).component();
-
-	const setupIRdescription2 = view.modelBuilder.text().withProps({
-		value: constants.SERVICE_CONTAINER_DESCRIPTION2,
-		CSSStyles: {
-			...styles.BODY_CSS
-		}
-	}).component();
-
-	const irSetupStep1Text = view.modelBuilder.text().withProps({
-		value: constants.SERVICE_STEP1,
-		CSSStyles: {
-			...styles.BODY_CSS
-		},
-		links: [
-			{
-				text: constants.SERVICE_STEP1_LINK,
-				url: 'https://aka.ms/sql-migration-shir-download'
-			}
-		]
-	}).component();
-
-	const irSetupStep2Text = view.modelBuilder.text().withProps({
-		value: constants.SERVICE_STEP2,
-		CSSStyles: {
-			...styles.BODY_CSS
-		}
-	}).component();
-
-	const irSetupStep3Text = view.modelBuilder.text().withProps({
-		value: constants.SERVICE_STEP3(testConnectionButton),
-		CSSStyles: {
-			'margin-top': '10px',
-			'margin-bottom': '10px',
-			...styles.BODY_CSS
-		}
-	}).component();
-
-	return view.modelBuilder.flexContainer().withItems(
-		[
-			setupIRHeadingText,
-			setupIRdescription1,
-			setupIRdescription2,
-			irSetupStep1Text,
-			irSetupStep2Text,
-			irSetupStep3Text,
-		]
-	).withLayout({
-		flexFlow: 'column'
-	}).component();
 }
